+++ ecosystem/typescript/sdk/src/plugins/index.ts
@@ -1,1 +1,3 @@
 export * from "./token_client";
+export * from "./coin_client";
+export * from "./faucet_client";
