+++ ecosystem/typescript/sdk/src/providers/index.ts
@@ -1,4 +1,3 @@
 export * from "./indexer";
 export * from "./aptos_client";
-export * from "./faucet_client";
 export * from "./provider";
