+++ ecosystem/typescript/sdk/src/tests/unit/aptos_account.test.ts
@@ -1,8 +1,8 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-import { AptosAccount, AptosAccountObject, getAddressFromAccountOrAddress } from "../../account/aptos_account";
-import { HexString } from "../../hex_string";
+import { AptosAccount, AptosAccountObject, getAddressFromAccountOrAddress } from "../../account";
+import { HexString } from "../../utils";
 
 const aptosAccountObject: AptosAccountObject = {
   address: "0x978c213990c4833df71548df7ce49d54c759d6b6d932de22b24d56060b7af2aa",
