+++ ecosystem/typescript/sdk/src/tests/unit/builder_utils.test.ts
@@ -1,7 +1,7 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-import { HexString } from "../../hex_string";
+import { HexString } from "../../utils";
 import {
   AccountAddress,
   Identifier,
