+++ ecosystem/typescript/sdk/src/tests/unit/hex_string.test.ts
@@ -1,7 +1,7 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-import { HexString } from "../../hex_string";
+import { HexString } from "../../utils";
 
 const withoutPrefix = "007711b4d0";
 const withPrefix = `0x${withoutPrefix}`;
