+++ ecosystem/typescript/sdk/src/tests/unit/transaction_vector.test.ts
@@ -40,7 +40,7 @@ import {
   TransactionArgumentU8Vector,
   TransactionArgumentU128,
 } from "../../aptos_types";
-import { HexString } from "../../hex_string";
+import { HexString } from "../../utils";
 import { TransactionBuilderEd25519 } from "../../transaction_builder/builder";
 
 // eslint-disable-next-line operator-linebreak
