+++ ecosystem/typescript/sdk/src/utils/index.ts
@@ -2,3 +2,4 @@ export * from "./misc";
 export * from "./memoize-decorator";
 export * from "./pagination_helpers";
 export * from "./api-endpoints";
+export * from "./hex_string";
